remove body code
