+++ all/bodies/__init__.py
@@ -1,4 +1,10 @@
-from .abstract import Body
-from .atari import DeepmindAtariBody, ParallelAtariBody
+from ._body import Body
+from .parallel import ParallelBody, ParallelFrameStack, ParallelRepeatActions
+from .rewards import RewardClipping
 
-__all__ = ["Body", "DeepmindAtariBody", "ParallelAtariBody"]
+__all__ = [
+    "Body",
+    "ParallelBody",
+    "ParallelFrameStack",
+    "ParallelRepeatActions",
+]
