+++ all/bodies/parallel_test.py
@@ -1,12 +1,9 @@
 import unittest
 import torch
 import torch_testing as tt
-import numpy as np
 from all.agents import Agent
-from all.environments import AtariEnvironment, State
 from all.bodies.parallel import ParallelRepeatActions
-from all.bodies import ParallelAtariBody
-
+from all.environments import State
 
 class MockAgent(Agent):
     def __init__(self, n, max_action=6):
@@ -65,28 +62,6 @@ class ParallelRepeatActionsTest(unittest.TestCase):
         for i, exp in enumerate(expected):
             self.assertEqual(actual[i], exp, msg=(("\nactual: %s\nexpected: %s")
                                                   % (actual, expected)))
-
-class ParallelAtariBodyTest(unittest.TestCase):
-    def test_runs(self):
-        np.random.seed(0)
-        torch.random.manual_seed(0)
-        n = 4
-        envs = []
-        for i in range(n):
-            env = AtariEnvironment('Breakout')
-            env.reset()
-            envs.append(env)
-        agent = MockAgent(n, max_action=4)
-        body = ParallelAtariBody(agent, envs, noop_max=30)
-
-        for _ in range(200):
-            states = [env.state for env in envs]
-            rewards = torch.tensor([env.reward for env in envs]).float()
-            actions = body.act(states, rewards)
-            for i, env in enumerate(envs):
-                if actions[i] is not None:
-                    env.step(actions[i])
-
 
 if __name__ == '__main__':
     unittest.main()
